booktool

+++ bookcms/Application/Custom/Controller/PerformController.class.php
@@ -17,6 +17,9 @@ class PerformController extends Controller {
 	}
 	private function save_media($page_id, $update=false){
 		if($update) {
+			if($_POST['media-changed']=='false'){
+				return;
+			}
 			M('meta')->where("page_id='".$page_id."' AND meta_key='media'")->delete();
 		}
 		if($_POST['media-list']) {
@@ -24,7 +27,12 @@ class PerformController extends Controller {
 			$num = 0;
 			foreach($medias as $media){
 				if(strpos($media['url'], 'http://')===0 && $media['title'] != ''){
-					$val = array('url'=>$media['url'], 'title'=>$media['title'], 'type'=>'mp3');
+					if(strpos($media['url'], '.m3u8')!==false){
+						$type="m3u8a";
+					}else{
+						$type="mp3";
+					}
+					$val = array('url'=>$media['url'], 'title'=>$media['title'], 'type'=>$type);
 					$json = json_encode($val);
 					mc_add_meta($page_id, 'media', $json, 'media');
 					$num++;
@@ -37,15 +45,19 @@ class PerformController extends Controller {
 	}
 	private function save_links($page_id, $update=false){
 		if($update) {
+			if($_POST['links-changed']=='false'){
+				return;
+			}
 			M('meta')->where("page_id='".$page_id."' AND meta_key='link'")->delete();
 		}
+		$link_type = array('play','page','wxbuy','otherbuy');
 		if($_POST['pro-links']) {
 			$links = I('post.pro-links');
 			$num = 0;
 			foreach($links as $link){
 				if(strpos($link['url'], 'http://')===0 && $link['title'] != ''){
 					$type = $link['type'];
-					if($type != 'play' && $type != 'buy'){
+					if(!in_array($type, $link_type)){
 						$type = 'page';
 					}
 					$val = array('url'=>$link['url'], 'title'=>$link['title'], 'type'=>$type);
@@ -97,42 +109,76 @@ class PerformController extends Controller {
 		}
 	}
 	private function save_pro(){
-	 $page_title = I('post.title');
-	 if($page_title && $_POST['content'] && is_numeric($_POST['price'])) {
-	 	$page['title'] = $page_title;
-	 	$page['content'] = mc_magic_in(mc_str_replace_base64($_POST['content']));
-	 	$page['type'] = 'pro';
-	 	$page['date'] = strtotime("now");
-	 	$result = M('page')->data($page)->add();
-	 	if($result) {
-	 		if($_POST['fmimg']) {
-	 			foreach($_POST['fmimg'] as $val) {
-	 				mc_add_meta($result,'fmimg',mc_save_img_base64($val,true));
-	 			}
-	 		}
-	 		$this->save_proparms($result);
-	 		$this->save_links($result);
-	 		$this->save_buyurl($result);
-	 		$this->save_basic_meta($result, 'isbn', 'post.isbn');
-	 		$this->save_basic_meta($result, 'price', 'post.price');
-	 		$this->save_basic_meta($result, 'kucun', 'post.kucun');
-	 		$this->save_basic_meta($result, 'term', 'post.term');
-	 		$this->save_basic_meta($result, 'keywords', 'post.keywords');
-	 		$this->save_basic_meta($result, 'description', 'post.description');
-	 		mc_add_meta($result,'author',mc_user_id());
-	 		do_go('publish_pro_end',$result);
-	 		return array('ret'=>true,'msg'=>'发布成功');
-	 	} else {
-	 		return array('ret'=>false,'msg'=>'发布失败！');
-	 	}
-	 } else {
- 		return array('ret'=>false,'msg'=>'请填写标题和内容');
-	 }
+		$page_title = I('post.title');
+		if($page_title && $_POST['content'] && is_numeric($_POST['price'])) {
+			$page['title'] = $page_title;
+			$page['content'] = mc_magic_in(mc_str_replace_base64($_POST['content']));
+			$page['type'] = 'pro';
+			$page['date'] = strtotime("now");
+			$result = M('page')->data($page)->add();
+			if($result) {
+				if($_POST['fmimg']) {
+					foreach($_POST['fmimg'] as $val) {
+						mc_add_meta($result,'fmimg',mc_save_img_base64($val,true));
+					}
+				}
+				$this->save_proparms($result);
+				$this->save_links($result);
+				$this->save_buyurl($result);
+				$this->save_basic_meta($result, 'isbn', 'post.isbn');
+				$this->save_basic_meta($result, 'price', 'post.price');
+				$this->save_basic_meta($result, 'kucun', 'post.kucun');
+				$this->save_basic_meta($result, 'term', 'post.term');
+				$this->save_basic_meta($result, 'keywords', 'post.keywords');
+				$this->save_basic_meta($result, 'description', 'post.description');
+				mc_add_meta($result,'author',mc_user_id());
+				do_go('publish_pro_end',$result);
+				return array('ret'=>true,'msg'=>$result);
+			} else {
+				return array('ret'=>false,'msg'=>'发布失败！');
+			}
+		} else {
+			return array('ret'=>false,'msg'=>'请填写标题和内容');
+		}
+	}
+	private function update_pro($page_id, $page_title){
+		if($_POST['fmimg']) {
+			mc_delete_meta($page_id,'fmimg');
+			foreach($_POST['fmimg'] as $val) {
+				mc_add_meta($page_id,'fmimg',mc_save_img_base64($val,true));
+			}
+		} else {
+			return array('ret'=>false,'msg'=>'请设置商品图片！');
+		}
+		$this->save_proparms($page_id, true);
+		$this->save_links($page_id, true);
+		$this->save_buyurl($page_id, true);
+		$this->save_basic_meta($page_id, 'isbn', 'post.isbn', true);
+		$this->save_basic_meta($page_id, 'price', 'post.price', true);
+		$this->save_basic_meta($page_id, 'kucun', 'post.kucun', true);
+		$this->save_basic_meta($page_id, 'term', 'post.term', true);
+		$this->save_basic_meta($page_id, 'keywords', 'post.keywords', true);
+		$this->save_basic_meta($page_id, 'description', 'post.description', true);
+		$page['title'] = $page_title;
+		$page['content'] = mc_magic_in(mc_str_replace_base64($_POST['content']));
+		M('page')->where(array('id'=>$page_id))->save($page);
+		return array('ret'=>true,'msg'=>'更新成功');
 	}
 
 	public function publish_pro_ajax(){
     	if(mc_is_admin() || mc_is_bianji()) {
-    	 	$result = $this->save_pro();
+    		$isbn = I('post.isbn');
+    		$page_id = M('meta')->where(array('meta_key'=>'isbn', 'meta_value'=>$isbn))->getField('page_id');
+    		if($page_id){
+    			$page_title = I('post.title');
+    			if($page_title && $_POST['content']){
+    				$result = $this->update_pro($page_id, $page_title);
+    			}else{
+    				$result= array('ret'=>false,'msg'=>'请填写标题和内容');
+    			}
+    		}else{
+    			$result = $this->save_pro();
+    		}
 	    } else {
 		    $result = array('ret'=>false,'msg'=>'没有权限');
 	    }
@@ -142,10 +188,10 @@ class PerformController extends Controller {
 	public function publish_pro(){
     	if(mc_is_admin() || mc_is_bianji()) {
     	 	$result = $this->save_pro();
-    	 	if($result.ret){
-    	 		$this->success('发布成功',U('pro/index/single?id='.$result));
+    	 	if($result['ret']){
+    	 		$this->success('发布成功',U('pro/index/single?id='.$result['msg']));
     	 	}else{
-    	 		$this->error(result.msg);
+    	 		$this->error($result['msg']);
     	 	}
 	    } else {
 		    $this->error('哥们，你放弃治疗了吗?',U('home/index/index'));
@@ -197,25 +243,14 @@ class PerformController extends Controller {
 	    	if($page_title && $_POST['content'] && is_numeric($_POST['id'])) {
     			$page_type = mc_get_page_field($_POST['id'],'type');
 	    		if($page_type=='pro') {
-		    		if($_POST['fmimg']) {
-		    			mc_delete_meta($_POST['id'],'fmimg');
-		    			foreach($_POST['fmimg'] as $val) {
-		    				mc_add_meta($_POST['id'],'fmimg',mc_save_img_base64($val,true));
-		    			}
-		    		} else {
-		    			$this->error('请设置商品图片！');
-		    		}
-	    			$this->save_proparms($_POST['id'], true);
-		    		$this->save_links($_POST['id'], true);
-		    		$this->save_buyurl($_POST['id'], true);
-	    			$this->save_basic_meta($_POST['id'], 'isbn', 'post.isbn', true);
-		    		$this->save_basic_meta($_POST['id'], 'price', 'post.price', true);
-	    			$this->save_basic_meta($_POST['id'], 'kucun', 'post.kucun', true);
-	    			$this->save_basic_meta($_POST['id'], 'term', 'post.term', true);
-	    			$this->save_basic_meta($_POST['id'], 'keywords', 'post.keywords', true);
-	    			$this->save_basic_meta($_POST['id'], 'description', 'post.description', true);
+	    			$result = $this->update_pro($_POST['id'],$page_title);
+	    			if(!$result['ret']){
+	    				$this->error($result['msg']);
+	    			}
 	    		} elseif ($page_type=='article') {
-	    			mc_update_meta($_POST['id'],'fmimg',mc_magic_in(mc_save_img_base64($_POST['fmimg'])));
+	    			if($_POST['fmimg']){
+	    				mc_update_meta($_POST['id'],'fmimg',mc_magic_in(mc_save_img_base64($_POST['fmimg'])));
+	    			}
 	    			$tags_str = I('post.tags');
 		    		if($tags_str) {
 			    		mc_delete_meta($_POST['id'],'tag');
@@ -229,10 +264,10 @@ class PerformController extends Controller {
 	    			$this->save_links($_POST['id'], true);
 	    			$this->save_media($_POST['id'], true);
 		    		$this->save_basic_meta($_POST['id'], 'term', 'post.term', true);;
+		    		$page['title'] = $page_title;
+		    		$page['content'] = mc_magic_in(mc_str_replace_base64($_POST['content']));
+		    		M('page')->where(array('id'=>$_POST['id']))->save($page);
 	    		}
-	    		$page['title'] = $page_title;
-	    		$page['content'] = mc_magic_in(mc_str_replace_base64($_POST['content']));
-	    		M('page')->where("id='".$_POST['id']."'")->save($page);
 	    		if($page_type=='pro') {
 		        	$this->success('编辑成功',U('pro/index/single?id='.$_POST['id']));
 	        	} elseif($page_type=='article') {
